add interface FormatReport class
